chore: replace command-handler.js with TypeScript version command-handler.ts
